Ensure the real path for the shim is resolved
Unfortunately this is a clasically hard problem on OSX. There's no
way (in shell) to give the OS some path on disk and say "make it
an absolute path and resolve all symlinks".

The OS has the system call (in C): man 3 realpath

But it doesn't expose it to the shell.

Python has a binding to that syscall so that's what we're doing here.
Would love to know of a better solution.

Also backticks can't be nested so it's best to avoid them in favor
of the $() construct.